refactor(fs): split out into utils-fs package
