+++ packages/@ionic/cli-framework/package.json
@@ -28,29 +28,23 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@ionic/utils-fs": "0.0.0",
     "@types/cross-spawn": "^6.0.0",
     "@types/debug": "0.0.30",
     "@types/inquirer": "0.0.43",
-    "@types/klaw": "^2.1.1",
     "@types/lodash": "^4.14.104",
-    "@types/make-dir": "^1.0.3",
     "@types/minimist": "^1.2.0",
-    "@types/ncp": "^2.0.1",
     "@types/node": "^6.0.101",
     "@types/rimraf": "^2.0.2",
     "@types/string-width": "^2.0.0",
     "@types/strip-ansi": "^3.0.0",
     "@types/untildify": "^3.0.0",
     "@types/wrap-ansi": "^3.0.0",
-    "@types/write-file-atomic": "^2.1.1",
     "chalk": "^2.3.0",
     "debug": "^3.1.0",
     "inquirer": "^6.0.0",
-    "klaw": "^3.0.0",
     "lodash": "^4.17.5",
-    "make-dir": "^1.3.0",
     "minimist": "^1.2.0",
-    "ncp": "^2.0.0",
     "rimraf": "^2.6.2",
     "slice-ansi": "^1.0.0",
     "stream-combiner2": "^1.1.1",
@@ -59,8 +53,7 @@
     "tree-kill": "1.2.0",
     "tslib": "^1.9.0",
     "untildify": "^3.0.2",
-    "wrap-ansi": "^4.0.0",
-    "write-file-atomic": "^2.3.0"
+    "wrap-ansi": "^4.0.0"
   },
   "devDependencies": {
     "jest": "^23.0.1",
