+++ packages/@ionic/cli-utils/package.json
@@ -48,6 +48,7 @@
   "dependencies": {
     "@ionic/cli-framework": "1.0.6",
     "@ionic/discover": "1.0.4",
+    "@ionic/utils-fs": "0.0.0",
     "@ionic/utils-network": "0.0.0",
     "chalk": "^2.3.0",
     "debug": "^3.1.0",
