+++ packages/@ionic/cli-utils/src/lib/integrations/cordova/__tests__/android.ts
@@ -11,7 +11,7 @@ describe('@ionic/cli-utils', () => {
       });
 
       it('should return undefined for a missing source.properties file', async () => {
-        jest.mock('@ionic/cli-framework/utils/fs', () => ({
+        jest.mock('@ionic/utils-fs', () => ({
           fsReadFile: () => {
             const err = new Error();
             err.code = 'ENOENT';
