+++ packages/@ionic/cli-utils/src/lib/integrations/cordova/__tests__/project.ts
@@ -1,4 +1,4 @@
-import * as fsSpy from '@ionic/cli-framework/utils/fs';
+import * as fsSpy from '@ionic/utils-fs';
 import * as project from '../project';
 
 describe('@ionic/cli-utils', () => {
