+++ packages/@ionic/cli-utils/src/lib/integrations/cordova/config.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 
 import { prettyPath } from '@ionic/cli-framework/utils/format';
-import { fsReadFile, fsWriteFile } from '@ionic/cli-framework/utils/fs';
+import { fsReadFile, fsWriteFile } from '@ionic/utils-fs';
 import chalk from 'chalk';
 import * as Debug from 'debug';
 import * as et from 'elementtree';
