+++ packages/@ionic/cli-utils/src/lib/serve.ts
@@ -1,7 +1,7 @@
 import { BaseError, LOGGER_LEVELS, OptionGroup, PromptModule, createPrefixedFormatter } from '@ionic/cli-framework';
-import { fsReadJsonFile } from '@ionic/cli-framework/utils/fs';
 import { killProcessTree, onBeforeExit, processExit } from '@ionic/cli-framework/utils/process';
 import { str2num } from '@ionic/cli-framework/utils/string';
+import { fsReadJsonFile } from '@ionic/utils-fs';
 import { NetworkInterface, findClosestOpenPort, getExternalIPv4Interfaces, isHostConnectable } from '@ionic/utils-network';
 import chalk from 'chalk';
 import * as Debug from 'debug';
