+++ packages/@ionic/cli-utils/src/lib/ssh-config.ts
@@ -1,7 +1,7 @@
 import * as os from 'os';
 import * as path from 'path';
 
-import { fileToString } from '@ionic/cli-framework/utils/fs';
+import { fileToString } from '@ionic/utils-fs';
 import * as SSHConfig from 'ssh-config';
 
 export { SSHConfig };
