+++ packages/@ionic/utils-fs/README.md
@@ -1,0 +1,1 @@
+# @ionic/utils-fs
