+++ packages/@ionic/utils-fs/package.json
@@ -1,19 +1,14 @@
 {
-  "name": "@ionic/lab",
-  "version": "1.0.8",
-  "description": "Ionic Lab utility for developing Ionic apps, used by Ionic CLI",
-  "bin": {
-    "ionic-lab": "./bin/ionic-lab"
-  },
+  "name": "@ionic/utils-fs",
+  "version": "0.0.0",
+  "description": "Filesystem utils for Node",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "homepage": "https://ionicframework.com/",
   "author": "Ionic Team <hi@ionic.io> (https://ionic.io)",
   "license": "MIT",
   "files": [
-    "bin/",
     "dist/",
-    "www/",
     "LICENSE",
     "README.md"
   ],
@@ -25,16 +20,11 @@
     "url": "https://github.com/ionic-team/ionic-cli/issues"
   },
   "scripts": {
-    "clean": "npm run clean.tsc",
-    "clean.tsc": "rimraf dist",
-    "clean.stencil": "rimraf www",
+    "clean": "rimraf dist",
     "lint": "tslint --config tslint.js --project tsconfig.json",
-    "build": "npm run build.tsc && npm run build.stencil",
-    "build.stencil": "npm run clean.stencil && stencil build",
-    "build.tsc": "npm run clean.tsc && tsc",
-    "watch": "npm run watch.tsc",
-    "watch.stencil": "npm run clean.stencil && concurrently \"stencil build --dev --watch\" \"stencil-dev-server --address localhost --httpPort 8900 --no-open\"",
-    "watch.tsc": "tsc -w --preserveWatchOutput",
+    "build": "npm run clean && tsc",
+    "watch": "tsc -w --preserveWatchOutput",
+    "test": "jest",
     "prepublishOnly": "npm run build",
     "precommit": "lint-staged"
   },
@@ -47,18 +37,20 @@
     ]
   },
   "dependencies": {
-    "@ionic/cli-framework": "1.0.6",
-    "@ionic/utils-fs": "0.0.0",
-    "chalk": "^2.3.0",
-    "express": "^4.16.2",
-    "tslib": "^1.9.0"
+    "debug": "^3.1.0",
+    "klaw": "^3.0.0",
+    "make-dir": "^1.3.0",
+    "ncp": "^2.0.0",
+    "tslib": "^1.9.0",
+    "write-file-atomic": "^2.3.0"
   },
   "devDependencies": {
-    "@stencil/core": "^0.12.4",
-    "@stencil/dev-server": "0.0.18",
-    "@types/express": "^4.11.0",
+    "@types/debug": "0.0.30",
+    "@types/klaw": "^2.1.1",
+    "@types/make-dir": "^1.0.3",
+    "@types/ncp": "^2.0.1",
     "@types/node": "^6.0.101",
-    "concurrently": "^4.0.1",
+    "@types/write-file-atomic": "^2.1.1",
     "jest": "^23.0.1",
     "jest-cli": "^23.0.1",
     "lint-staged": "^7.2.0",
