+++ packages/@ionic/utils-network/package.json
@@ -37,7 +37,8 @@
     ]
   },
   "dependencies": {
-    "debug": "^3.1.0"
+    "debug": "^3.1.0",
+    "tslib": "^1.9.0"
   },
   "devDependencies": {
     "@types/debug": "0.0.30",
