+++ packages/@ionic/v1-toolkit/package.json
@@ -43,6 +43,7 @@
   },
   "dependencies": {
     "@ionic/cli-framework": "1.0.6",
+    "@ionic/utils-fs": "0.0.0",
     "chalk": "^2.3.0",
     "chokidar": "^2.0.0",
     "debug": "^3.1.0",
