+++ packages/cli-scripts/package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "@ionic/cli-framework": "1.0.6",
     "@ionic/cli-utils": "2.2.0",
+    "@ionic/utils-fs": "0.0.0",
     "ansi-styles": "^3.2.0",
     "chalk": "^2.3.0",
     "escape-string-regexp": "^1.0.5",
