+++ packages/ionic/package.json
@@ -55,6 +55,7 @@
   "dependencies": {
     "@ionic/cli-framework": "1.0.6",
     "@ionic/cli-utils": "2.2.0",
+    "@ionic/utils-fs": "0.0.0",
     "chalk": "^2.3.0",
     "debug": "^3.1.0",
     "lodash": "^4.17.5",
